requirements.txt
Flask==1.1.1
Flask-SQLAlchemy>=2.5.0